+++ git_services/git_services/sidecar/config.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+import dataconf
+
+from git_services.cli.sentry import SentryConfig
+
+
+@dataclass
+class Config:
+    sentry: SentryConfig
+
+
+def config_from_env() -> Config:
+    return dataconf.env("GIT_RPC_", Config)