+++ renku_notebooks/api/classes/server.py
@@ -2,7 +2,7 @@ from flask import current_app
 import gitlab
 from itertools import chain
 from kubernetes import client
-from kubernetes.client.rest import ApiException
+from kubernetes.client.exceptions import ApiException
 from kubernetes.client.models import V1DeleteOptions
 import base64
 import json
@@ -10,7 +10,6 @@ from urllib.parse import urlparse, urljoin
 
 
 from ..amalthea_patches import (
-    autosave as autosave_patches,
     general as general_patches,
     git_proxy as git_proxy_patches,
     git_sidecar as git_sidecar_patches,
@@ -283,13 +282,13 @@ class UserServer:
                 general_patches.session_tolerations(),
                 general_patches.session_affinity(),
                 general_patches.session_node_selector(),
+                general_patches.termination_grace_period(),
                 jupyter_server_patches.args(),
                 jupyter_server_patches.env(self),
                 jupyter_server_patches.image_pull_secret(self),
                 jupyter_server_patches.disable_service_links(),
-                autosave_patches.main(),
                 git_proxy_patches.main(self),
-                git_sidecar_patches.main(),
+                git_sidecar_patches.main(self),
                 general_patches.oidc_unverified_email(self),
                 cloudstorage_patches.main(self),
                 # init container for certs must come before all other init containers
@@ -475,24 +474,38 @@ class UserServer:
         else:
             return status
 
-    def get_logs(self, max_log_lines=0, container_name="jupyter-server"):
-        """Get the logs of the k8s pod that runs the user server."""
+    def get_logs(self, max_log_lines=0):
+        """Get the logs of all containers in the server pod."""
         js = self.js
         if js is None:
             return None
+        output = {}
         pod_name = js["status"]["mainPod"]["name"]
-        if max_log_lines == 0:
-            logs = self._k8s_client.read_namespaced_pod_log(
-                pod_name, self._k8s_namespace, container=container_name
-            )
+        all_containers = js["status"]["mainPod"]["status"].get(
+            "containerStatuses", []
+        ) + js["status"]["mainPod"]["status"].get("initContainerStatuses", [])
+        for container in all_containers:
+            container_name = container["name"]
+            try:
+                logs = self._k8s_client.read_namespaced_pod_log(
+                    pod_name,
+                    self._k8s_namespace,
+                    container=container_name,
+                    tail_lines=max_log_lines if max_log_lines > 0 else None,
+                    timestamps=True,
+                )
+            except ApiException as err:
+                if err.status in [400, 404]:
+                    continue  # container does not exist or is not ready yet
+                else:
+                    raise
+            else:
+                output[container_name] = logs
+
+        if len(output.keys()) == 0:
+            return None
         else:
-            logs = self._k8s_client.read_namespaced_pod_log(
-                pod_name,
-                self._k8s_namespace,
-                tail_lines=max_log_lines,
-                container=container_name,
-            )
-        return logs
+            return output
 
     @property
     def server_url(self):