+++ renku_notebooks/api/schemas.py
@@ -642,23 +642,20 @@ class ServerOptionsUI(Schema):
     disk_request = fields.Nested(
         ServerOptionUIDisk(), required="disk_request" in config.SERVER_OPTIONS_UI.keys()
     )
-    # TODO: enable when the UI supports fully s3 buckets
-    # currently passing this breaks the sessions settings page
-    # cloudstorage = fields.Nested(CloudStorageServerOption(), required=True)
+    cloudstorage = fields.Nested(CloudStorageServerOption(), required=True)
 
 
-class ServerLogs(Schema):
-    """
-    The list of k8s logs (one log line per list element)
-    for the pod that runs the jupyter server.
-    """
+_ServerLogs = Schema.from_dict({"jupyter-server": fields.String(required=False)})
 
-    items = fields.List(fields.Str())
 
-    @post_dump
-    @post_load
-    def remove_item_key(self, data, **kwargs):
-        return data.get("items", [])
+class ServerLogs(_ServerLogs):
+    class Meta:
+        unknown = INCLUDE  # only affects loading, not dumping
+
+    @post_dump(pass_original=True)
+    def keep_unknowns(self, output, orig, **kwargs):
+        output = {**orig, **output}
+        return output
 
 
 def _in_range(value, value_range):